+++ import_export/admin.py
@@ -853,7 +853,12 @@ class ExportMixin(BaseExportMixin, ImportExportMixinBase):
         context["fields_list"] = [
             (
                 res.get_display_name(),
-                [field.column_name for field in res().get_user_visible_fields()],
+                [
+                    field.column_name
+                    for field in res(
+                        **self.get_export_resource_kwargs(request)
+                    ).get_user_visible_fields()
+                ],
             )
             for res in self.get_export_resource_classes()
         ]